reorganizes componentss into container/components folders
