+++ src/components/Avatar.js
